Restructure project for eslint and fix the errors
